Client refactor WIP
