+++ bin/maze-runner
@@ -12,6 +12,10 @@ require_relative '../lib/maze/bitbar_utils'
 require_relative '../lib/maze/bitbar_devices'
 require_relative '../lib/maze/bugsnag_config'
 require_relative '../lib/maze/capabilities'
+require_relative '../lib/maze/client/appium/base_client'
+require_relative '../lib/maze/client/appium/bb_client'
+require_relative '../lib/maze/client/appium/bs_client'
+require_relative '../lib/maze/client/appium/local_client'
 require_relative '../lib/maze/compare'
 require_relative '../lib/maze/docker'
 require_relative '../lib/maze/document_server'
