+++ lib/maze/client/appium/bs_client.rb
@@ -1,0 +1,11 @@
+module Maze
+  module Client
+    module Appium
+      class BrowserStackClient < BaseClient
+        class << self
+
+        end
+      end
+    end
+  end
+end
